remove bugs

+++ backend/controllers/product.controller.js
@@ -1,4 +1,5 @@
 import Product from "../models/product.model.js";
+import mongoose from "mongoose";
 
 export const getProducts = async (req, res) => {
   try {
@@ -20,18 +21,19 @@ export const updateProduct = async (req, res) => {
   const { id } = req.params;
   const productData = req.body;
 
+  // checking if id is valid
+  if (!id || !productData || !mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({
+      success: false,
+      message: "Please provide a valid product ID to update",
+    });
+  }
+
   try {
     const updatedProduct = await Product.findByIdAndUpdate(id, productData, {
       new: true,
       runValidators: true,
     });
-
-    if (!updatedProduct) {
-      return res.status(404).json({
-        success: false,
-        message: "Product not found",
-      });
-    }
 
     res.status(200).json({
       success: true,
@@ -74,6 +76,14 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  // checking if id is valid
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({
+      success: false,
+      message: "Please provide a valid product ID to delete",
+    });
+  }
 
   try {
     const product = await Product.findByIdAndDelete(id);
